Added strip.placement = "outside"
Using a secondary y-axis, facet_grid with the strip.background = ggplot2::element_blank() results in the label being between the border of the plot and the axis ticks, which looks strange. strip.placement = "outside" moves the axis to the plot and the label further outside resulting in a better appearance.